[feat] add sgx intergration
