+++ deploy/helm/templates/rbac.yaml
@@ -1,19 +1,48 @@
-# TokenReview Create ClusterRole
+
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   namespace: {{ .Values.installation.namespace }}
-  name: token-review-creator
+  name: polaris-security-cluster-role
 rules:
+# TokenReview creation
 - apiGroups: ["authentication.k8s.io"]
   resources: ["tokenreviews"]
   verbs: ["create"]
+{{ if .Values.polaris_security.signer }}
+# Kubernetes CSR Usage ClusterRole when using another singer
+- apiGroups:
+    - certificates.k8s.io
+  resources:
+    - certificatesigningrequests
+  verbs:
+    - delete
+    - deletecollection
+    - get
+    - put
+    - list
+    - create
+- apiGroups:
+    - certificates.k8s.io
+  resources:
+    - certificatesigningrequests/approval
+  verbs:
+    - update
+- apiGroups:
+  - certificates.k8s.io
+  resources:
+  - signers
+  resourceNames:
+  - {{ .Values.polaris_security.signer }}
+  verbs:
+  - approve
+{{ end }}
 ---
 # bind the role to polaris-security
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: create-token-review
+  name: polaris-security-role-binding
   namespace: {{ .Values.installation.namespace }}
 subjects:
 - kind: ServiceAccount
@@ -21,5 +50,5 @@ subjects:
   namespace: {{ .Values.installation.namespace }}
 roleRef:
   kind: ClusterRole
-  name: token-review-creator
+  name: polaris-security-cluster-role
   apiGroup: rbac.authorization.k8s.io
