+++ deploy/helm/values.yaml
@@ -2,10 +2,12 @@ polaris_security:
   containers:
       image: polarismesh/polaris-security
       tag: latest
-      imagePullPolicy: Never
+      imagePullPolicy: IfNotPresent
   bind: 0.0.0.0
   port: 8888
   default_cert_ttl: 3600
   dns_names: polaris-security,polaris-security.polaris-system.svc,polaris-security.polaris-system.svc.cluster.local
+  #signer: tcsissuer.tcs.intel.com/sandbox.my-ca
+  signer: ""
 installation:
   namespace: polaris-system
