+++ deploy/load-certs-into-k8s.sh
@@ -14,9 +14,12 @@ function GenerateDefaultMaterials() {
     rm -rf ${TMP_DIR}
 }
 
+NAMESPACE=polaris-system
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]:-$0}"; )" &> /dev/null && pwd 2> /dev/null; )";
 TOOL_DIR=${SCRIPT_DIR}/../tools
 CERTS_DIR=${SCRIPT_DIR}/../certs
+
+mkdir -p ${CERTS_DIR}
 
 if [[ -z ${CA_CERT_PATH} ]] || [[ -z ${CA_KEY_PATH} ]] || [[ -z ${ROOT_CERT_PATH} ]] || [[ -z ${CERT_CHAIN_PATH} ]]; then
     GenerateDefaultMaterials
@@ -27,13 +30,15 @@ if [[ -z ${CA_CERT_PATH} ]] || [[ -z ${CA_KEY_PATH} ]] || [[ -z ${ROOT_CERT_PATH
     CERT_CHAIN_PATH=${CERTS_DIR}/cert-chain.pem
 fi
 
+kubectl create ns ${NAMESPACE}
+
 # create secret for polaris-security
-kubectl create secret generic polaris-security-secret -n polaris-system \
+kubectl create secret generic polaris-security-secret -n ${NAMESPACE} \
     --from-file=ca-cert=${CA_CERT_PATH} \
     --from-file=ca-key=${CA_KEY_PATH} \
     --from-file=root-cert=${ROOT_CERT_PATH} \
     --from-file=cert-chain=${CERT_CHAIN_PATH}
 
 # create secret for polaris-sidecar
-kubectl create secret generic polaris-sidecar-secret -n polaris-system \
+kubectl create secret generic polaris-sidecar-secret -n ${NAMESPACE} \
     --from-file=root-cert=${ROOT_CERT_PATH}
